+++ src/app/dataset-flow/dataset-flow-details/dataset-flow-details.component.spec.ts
@@ -26,6 +26,8 @@ import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 import { FlowDetailsTabs } from "./dataset-flow-details.types";
 import ProjectLinks from "src/app/project-links";
 import { mockDatasetFlowByIdResponse, mockFlowSummaryDataFragments } from "src/app/api/mock/dataset-flow.mock";
+import { DataAccessPanelComponent } from "src/app/components/data-access-panel/data-access-panel.component";
+import { DataAccessPanelModule } from "src/app/components/data-access-panel/data-access-panel.module";
 
 describe("DatasetFlowDetailsComponent", () => {
     let component: DatasetFlowDetailsComponent;
@@ -46,6 +48,7 @@ describe("DatasetFlowDetailsComponent", () => {
                 SearchAdditionalButtonsComponent,
                 SearchAdditionalButtonsNavComponent,
                 DatasetViewMenuComponent,
+                DataAccessPanelComponent,
             ],
             providers: [
                 {
@@ -96,6 +99,7 @@ describe("DatasetFlowDetailsComponent", () => {
                 MatTabsModule,
                 FormsModule,
                 BrowserAnimationsModule,
+                DataAccessPanelModule,
             ],
         }).compileComponents();