+++ src/app/dataset-view/dataset-view-menu/dataset-view-menu.component.spec.ts
@@ -1,10 +1,9 @@
-import AppValues from "src/app/common/app.values";
 import { DatasetNavigationInterface } from "../dataset-view.interface";
 import { FormsModule } from "@angular/forms";
-import { ComponentFixture, fakeAsync, flush, TestBed, tick } from "@angular/core/testing";
+import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { MatMenuModule } from "@angular/material/menu";
 import { DatasetViewMenuComponent } from "./dataset-view-menu.component";
-import { emitClickOnElementByDataTestId, getElementByDataTestId } from "src/app/common/base-test.helpers.spec";
+import { emitClickOnElementByDataTestId } from "src/app/common/base-test.helpers.spec";
 import { BrowserAnimationsModule } from "@angular/platform-browser/animations";
 import { MatButtonToggleModule } from "@angular/material/button-toggle";
 import { AngularSvgIconModule } from "angular-svg-icon";
@@ -12,6 +11,10 @@ import { HttpClientTestingModule } from "@angular/common/http/testing";
 import { MatTabsModule } from "@angular/material/tabs";
 import { MatIconModule } from "@angular/material/icon";
 import { mockDatasetBasicsDerivedFragment, mockFullPowerDatasetPermissionsFragment } from "src/app/search/mock.data";
+import { DataAccessPanelComponent } from "src/app/components/data-access-panel/data-access-panel.component";
+import { MatCheckboxModule } from "@angular/material/checkbox";
+import { MatDividerModule } from "@angular/material/divider";
+import { MatTooltipModule } from "@angular/material/tooltip";
 
 describe("DatasetViewMenuComponent", () => {
     let component: DatasetViewMenuComponent;
@@ -39,8 +42,11 @@ describe("DatasetViewMenuComponent", () => {
                 MatIconModule,
                 AngularSvgIconModule.forRoot(),
                 HttpClientTestingModule,
+                MatDividerModule,
+                MatCheckboxModule,
+                MatTooltipModule,
             ],
-            declarations: [DatasetViewMenuComponent],
+            declarations: [DatasetViewMenuComponent, DataAccessPanelComponent],
         }).compileComponents();
 
         fixture = TestBed.createComponent(DatasetViewMenuComponent);
@@ -71,21 +77,4 @@ describe("DatasetViewMenuComponent", () => {
             }
         });
     });
-
-    it("should copy to clipboard", fakeAsync(() => {
-        emitClickOnElementByDataTestId(fixture, "searchAdditionalButtons");
-
-        const menu = getElementByDataTestId(fixture, "menu");
-        expect(menu).toBeDefined();
-
-        const copyToClipboardButton = getElementByDataTestId(fixture, "copyToClipboard");
-        emitClickOnElementByDataTestId(fixture, "copyToClipboard");
-        expect(copyToClipboardButton.classList.contains("clipboard-btn--success")).toEqual(true);
-
-        tick(AppValues.LONG_DELAY_MS);
-
-        expect(copyToClipboardButton.classList.contains("clipboard-btn--success")).toEqual(false);
-
-        flush();
-    }));
 });