+++ src/app/dataset-view/dataset-view-menu/dataset-view-menu.component.ts
@@ -11,8 +11,6 @@ import {
 } from "@angular/core";
 import { MatSidenav } from "@angular/material/sidenav";
 import { DatasetNavigationInterface, DatasetViewTypeEnum } from "../dataset-view.interface";
-import { Clipboard } from "@angular/cdk/clipboard";
-import AppValues from "../../common/app.values";
 import { SideNavHelper } from "../../common/sidenav.helper";
 import { isMobileView, promiseWithCatch } from "src/app/common/app.helpers";
 import { DatasetBasicsFragment, DatasetPermissionsFragment } from "src/app/api/kamu.graphql.interface";
@@ -35,13 +33,9 @@ export class DatasetViewMenuComponent implements OnInit, AfterViewInit {
     @Input() datasetViewType: DatasetViewTypeEnum;
     @Input() isMinimizeSearchAdditionalButtons: boolean;
 
-    public clipboardKamuCli = "";
-    public clipboardKafka = "";
-
     private sideNavHelper: SideNavHelper;
 
     constructor(
-        private clipboard: Clipboard,
         private datasetPermissionsServices: DatasetPermissionsService,
         private widgetHeightService: WidgetHeightService,
     ) {}
@@ -56,26 +50,6 @@ export class DatasetViewMenuComponent implements OnInit, AfterViewInit {
     public ngOnInit(): void {
         if (this.sidenav) {
             this.sideNavHelper = new SideNavHelper(this.sidenav);
-        }
-        this.initClipboardHints();
-    }
-
-    public copyToClipboard(event: MouseEvent, text: string): void {
-        this.clipboard.copy(text);
-
-        if (event.currentTarget !== null) {
-            const currentElement: HTMLButtonElement = event.currentTarget as HTMLButtonElement;
-            const currentElementChildren: HTMLCollectionOf<HTMLElement> =
-                currentElement.children as HTMLCollectionOf<HTMLElement>;
-            setTimeout(() => {
-                currentElementChildren[0].style.display = "inline-block";
-                currentElementChildren[1].style.display = "none";
-                currentElement.classList.remove("clipboard-btn--success");
-            }, AppValues.LONG_DELAY_MS);
-
-            currentElementChildren[0].style.display = "none";
-            currentElementChildren[1].style.display = "inline-block";
-            currentElement.classList.add("clipboard-btn--success");
         }
     }
 
@@ -164,10 +138,5 @@ export class DatasetViewMenuComponent implements OnInit, AfterViewInit {
                 this.datasetViewMenuComponent.nativeElement.offsetHeight +
                     this.datasetViewMenuComponent.nativeElement.offsetTop,
             );
-    }
-
-    private initClipboardHints(): void {
-        this.clipboardKamuCli = `kamu pull kamu.dev/${this.datasetBasics.alias}`;
-        this.clipboardKafka = `https://api.kamu.dev/kafka/${this.datasetBasics.alias}`;
     }
 }